additional marshalling tests and fixes

+++ errai-marshalling/src/main/java/org/jboss/errai/marshalling/server/marshallers/ServerShortMarshaller.java
@@ -36,7 +36,7 @@ public class ServerShortMarshaller extends AbstractNumberMarshaller<Object, Shor
   @Override
   public Short demarshall(Object o, MarshallingSession ctx) {
     if (o instanceof Map) {
-      return ((Short) ((Map) o).get(SerializationParts.NUMERIC_VALUE)).shortValue();
+      return ((Double) ((Map) o).get(SerializationParts.NUMERIC_VALUE)).shortValue();
     }
     else {
       return (Short) o;
